`optimize_arrow_method_prop` is pure
